# release-23.1: sql/stats: improve query to ensure all tables